refactor IssuesList
