+++ src/components/IssuesList.jsx
@@ -1,7 +1,8 @@
 // IssuesList.js
 import React from "react";
-import { Box, Chip, List, ListItem, Paper, Typography } from "@mui/material";
+import { Box, List, Paper, Typography } from "@mui/material";
 import { styled } from "@mui/material/styles";
+import IssueCard from "./IssueCard";
 
 const IssuesListContainer = styled(Paper)(({ theme }) => ({
   height: "100%",
@@ -23,37 +24,6 @@ const ScrollableList = styled(List)(({ theme }) => ({
   },
 }));
 
-const StyledListItem = styled(ListItem)(({ theme }) => ({
-  borderBottom: `1px solid ${theme.palette.divider}`,
-  cursor: "pointer",
-  transition: theme.transitions.create(["background-color"], {
-    duration: theme.transitions.duration.shorter,
-  }),
-  "&:hover": {
-    backgroundColor: theme.palette.action.hover,
-  },
-}));
-
-const IssueTitle = ({ title, type, getChipColor }) => (
-  <Typography component="div" variant="subtitle1" sx={{ mb: 1 }}>
-    <Box sx={{ display: "flex", justifyContent: "space-between" }}>
-      <span>{title}</span>
-      <Chip label={type} size="small" color={getChipColor(type)} variant="outlined" />
-    </Box>
-  </Typography>
-);
-
-const IssueDetails = ({ description, location, timestamp, formatTime }) => (
-  <Typography component="div" variant="body2" color="text.secondary">
-    <Box sx={{ mb: 1 }}>{description}</Box>
-    <Box sx={{ display: "flex", gap: 1, fontSize: "12px" }}>
-      <span>{location}</span>
-      <span>•</span>
-      <span>{formatTime(timestamp)}</span>
-    </Box>
-  </Typography>
-);
-
 const IssuesList = ({ issues, hoveredIssue, selectedIssue, handleIssueSelect, setHoveredIssue, getChipColor, formatTime }) => (
   <IssuesListContainer elevation={0}>
     <Box sx={{ p: 2, borderBottom: 1, borderColor: "divider" }}>
@@ -63,25 +33,17 @@ const IssuesList = ({ issues, hoveredIssue, selectedIssue, handleIssueSelect, se
     </Box>
     <ScrollableList>
       {issues.map((issue) => (
-        <StyledListItem
+        <IssueCard
           key={issue.id}
-          onMouseEnter={() => setHoveredIssue(issue.id)}
-          onMouseLeave={() => setHoveredIssue(null)}
-          onClick={() => handleIssueSelect(issue)}
-          sx={{
-            bgcolor: hoveredIssue === issue.id || selectedIssue === issue.id ? "action.hover" : "transparent",
-          }}
-        >
-          <Box sx={{ width: "100%" }}>
-            <IssueTitle title={issue.title} type={issue.type} getChipColor={getChipColor} />
-            <IssueDetails
-              description={issue.description}
-              location={issue.location}
-              timestamp={issue.timestamp}
-              formatTime={formatTime}
-            />
-          </Box>
-        </StyledListItem>
+          issue={issue}
+          isSelected={selectedIssue === issue.id}
+          isHovered={hoveredIssue === issue.id}
+          handleMouseEnter={setHoveredIssue}
+          handleMouseLeave={() => setHoveredIssue(null)}
+          handleSelect={handleIssueSelect}
+          getChipColor={getChipColor}
+          formatTime={formatTime}
+        />
       ))}
     </ScrollableList>
   </IssuesListContainer>
